requirements.txt
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.11.2
django-import-export==1.0.1
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
et-xmlfile==1.0.1
idna==2.7
jdcal==1.4
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.6
PyJWT==1.6.4
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
